Clean up, move loop to within getCamaerDate

+++ cweb/tasks.hpp
@@ -54,7 +54,7 @@ public:
 	void taskHunt();
 
     int  startCamera();
-    int  getCameraData(int socketOrAddr);
+    float  getCameraData(int socketOrAddr);
     int  stopCamera();
 };
